Merge duplicate subfolder name and fix acronyms
* Merge duplicare subfolder name
Usually API that requests a specific resource (by SID) are following API that requests
all available resources (see for example /Credentials/AWS and Credentials/AWS/:Sid)
In this case we group them together in the same folder

* Fox acronyms
The way we convert camel case into a sentence is to add
a space before a capital letter. That doesn't work for
acronyms (e.g. SMS). Change Regex to search for sequence of
capital letters (Quantifier +)